+++ lib/ruby-lokalise-api/client.rb
@@ -1,3 +1,4 @@
+require 'ruby-lokalise-api/rest/branches'
 require 'ruby-lokalise-api/rest/languages'
 require 'ruby-lokalise-api/rest/teams'
 require 'ruby-lokalise-api/rest/projects'